Replace 'true' assertion status by 'trust' when needed.
Note: currently 'true' is used only for the analysis output. In most of
the cases we should use 'trust' (e.g., for assertions that specify that
some predicate/property is native, defined via the foreign interface,
etc.) or no status (equivalent to 'check') for assertions that document
the interface of a predicate in a module.

Src-commit: b23872a343f99448f721a23975dc7b332c53cdf9